+++ src/SMAPI/ISemanticVersion.cs
@@ -24,6 +24,9 @@ namespace StardewModdingAPI
         /*********
         ** Accessors
         *********/
+        /// <summary>Whether this is a pre-release version.</summary>
+        bool IsPrerelease();
+
         /// <summary>Get whether this version is older than the specified version.</summary>
         /// <param name="other">The version to compare with this instance.</param>
         bool IsOlderThan(ISemanticVersion other);